Changes suggested by include-what-you-use
This commit applies changes suggested by include-what-you-use.  I did
not apply all of them as there are some false positives/negatives and
it did not recognize forward decls in some cases, and also tends to
suggest headers that should not be included, such as files in the
bits/ directory on Linux, but most changes have been applied.